api/v1: remove write-only NodeAddressingElement.AddressType property
This field is only ever set but never read (i.e. inside the CNI).
It is fine to make this change because the agent and CNI are always
deployed from the same Cilium version in lockstep, i.e. with matching
API. No other API clients use NodeAddressingElement or NodeAddressing
information. Because the field is never read, it shouldn't even cause
problems with CNI plugin using an older version of the API.

This also allows to drop pkg/node.getNodeIP and fold it into
pkg/node.GetNodeIP.

Signed-off-by: Tobias Klauser <tobias@cilium.io>